fix(v3.6.0): address CR review — FK types, security hardening, sliding window
Schema fixes (root cause of MySQL/MariaDB CI failure):
- migrations.php, schema.mysql.sql: INT UNSIGNED → BIGINT UNSIGNED for
  totp_backup_codes.id and user_id to match users.id (BIGINT UNSIGNED)
- migrations.php, schema.pgsql.sql: SERIAL/INTEGER → BIGINT GENERATED BY
  DEFAULT AS IDENTITY / BIGINT for pgsql parity

Security fixes:
- lib.php: ipam_totp_encrypt_secret/decrypt_secret — fail closed (RuntimeException)
  when app_secret is empty; prevents fixed-key encryption on misconfigured installs
- lib.php: ipam_totp_verify_backup_code — atomic UPDATE WHERE used_at IS NULL
  + rowCount check to prevent race-condition double-use of backup codes
- lib.php: ipam_api_key_rate_limit_check — true sliding-window approximation
  using two-bucket weighted count instead of fixed-window; fixes ~2x burst boundary
- api.php: Retry-After now returns seconds until end of current window
  (windowSec - time() % windowSec) instead of full window duration
- lib.php: ipam_session_enforce_absolute_lifetime — treat absolute_lifetime_minutes
  <= 0 as disabled; don't seed _abs_expires pre-auth (moved to login_user())
- lib.php: login_user() now seeds _abs_expires immediately after authentication
- login.php: store only totp_pending_uid in session (not username/role);
  add autocomplete=username / autocomplete=current-password to login inputs
- totp_verify.php: reload fresh username/role/totp_enabled from DB on every POST;
  reject missing/inactive user rows instead of falling through to login_user();
  call ipam_clear_persistent_lockout() on successful 2FA to reset failure counter

Docs:
- CLAUDE.md: fix stale v3.5.0 reference in migration-policy preface → v3.6.0
- testing/playwright/tests/reverse-proxy.spec.ts: assert autocomplete attributes
  (username, current-password) not just field visibility

+++ Simple-PHP-IPAM/schema.mysql.sql
@@ -593,8 +593,8 @@ CREATE TABLE IF NOT EXISTS schema_migrations (
 -- totp_backup_codes (v3.6.0, #418)
 -- ---------------------------------------------------------------------------
 CREATE TABLE IF NOT EXISTS totp_backup_codes (
-  id          INT UNSIGNED NOT NULL AUTO_INCREMENT,
-  user_id     INT UNSIGNED NOT NULL,
+  id          BIGINT UNSIGNED NOT NULL AUTO_INCREMENT,
+  user_id     BIGINT UNSIGNED NOT NULL,
   code_hash   TEXT NOT NULL,
   used_at     DATETIME,
   PRIMARY KEY (id),
